requirements.txt
streamlit==1.12.0
gtts==2.3.2
pillow==9.2.0
pytesseract==0.3.10
tesseract==0.1.3